move parse_args in main to Dto
